refactor(eval): share rule authoring primitives
Extract deterministic profiling, value-form inference, and YAML emission so tuning can reuse the proven drafting behavior without duplicating it.
